add authorization in routes

+++ ApiCompras.Api/Controllers/PersonController.cs
@@ -1,27 +1,37 @@
 ﻿using ApiCompras.Application.DTOs;
 using ApiCompras.Application.Services;
 using ApiCompras.Application.Services.Interface;
+using ApiCompras.Domain.Authentication;
 using ApiCompras.Domain.FiltersDb;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
 namespace ApiCompras.Api.Controllers
 {
-    [Authorize]
     [ApiController]
     [Route("v1/person")]
-    public class PersonController : ControllerBase
+    public class PersonController : BaseController
     {
         private readonly IPersonService _personService;
-
-        public PersonController(IPersonService personService)
+        private readonly ICurrentUser _currentUser;
+        private List<string> _permissionNeeded = new List<string>() { "Admin" };
+        private readonly List<string> _permissionUser;
+        public PersonController(
+            IPersonService personService,
+            ICurrentUser currentUser)
         {
             _personService = personService;
+            _currentUser = currentUser;
+            _permissionUser = _currentUser.Permissions.Split(",").ToList() ?? new List<string>();
         }
 
         [HttpGet]
-        public async Task<ActionResult<ResultService<List<PersonDTO>>>> GetAsync()
+        public async Task<IActionResult> GetAsync()
         {
+            _permissionNeeded.Add("BuscaPessoa");
+            if (!ValidatePermission(_permissionUser, _permissionNeeded))
+                return Forbidden();
+
             var result = await _personService.GetAsync();
 
             if (!result.IsSuccess)
@@ -31,9 +41,13 @@ namespace ApiCompras.Api.Controllers
         }
 
         [HttpGet("{id}")]
-        public async Task<ActionResult<ResultService<PersonDTO>>> GetByIdAsync(
+        public async Task<IActionResult> GetByIdAsync(
             [FromRoute] int id)
         {
+            _permissionNeeded.Add("BuscaPessoa");
+            if (!ValidatePermission(_permissionUser, _permissionNeeded))
+                return Forbidden();
+
             var result = await _personService.GetByIdAsync(id);
 
             if (!result.IsSuccess)
@@ -43,17 +57,25 @@ namespace ApiCompras.Api.Controllers
         }
 
         [HttpGet("paged")]
-        public async Task<ActionResult> GetPagedAsync(
-            [FromQuery]PersonFilterDb filter)
+        public async Task<IActionResult> GetPagedAsync(
+            [FromQuery] PersonFilterDb filter)
         {
+            _permissionNeeded.Add("BuscaPessoa");
+            if (!ValidatePermission(_permissionUser, _permissionNeeded))
+                return Forbidden();
+
             var result = await _personService.GetPagedAsync(filter);
             return Ok(result);
         }
 
         [HttpPost]
-        public async Task<ActionResult<ResultService<PersonDTO>>> CreateAsync(
+        public async Task<IActionResult> CreateAsync(
             [FromBody] PersonDTO person)
         {
+            _permissionNeeded.Add("CadastroPessoa");
+            if (!ValidatePermission(_permissionUser, _permissionNeeded))
+                return Forbidden();
+
             var result = await _personService.CreateAsync(person);
             if (result.IsSuccess)
                 return Ok(result);
@@ -62,9 +84,13 @@ namespace ApiCompras.Api.Controllers
         }
 
         [HttpPut]
-        public async Task<ActionResult<ResultService<PersonDTO>>> UpddateAsync(
-            [FromBody]PersonDTO person)
+        public async Task<IActionResult> UpddateAsync(
+            [FromBody] PersonDTO person)
         {
+            _permissionNeeded.Add("EditaPessoa");
+            if (!ValidatePermission(_permissionUser, _permissionNeeded))
+                return Forbidden();
+
             var result = await _personService.UpdateAsync(person);
             if (result.IsSuccess)
                 return Ok(result);
@@ -73,9 +99,12 @@ namespace ApiCompras.Api.Controllers
         }
 
         [HttpDelete("{id}")]
-        public async Task<ActionResult<ResultService>> DeleteAsync(
-            [FromRoute]int id)
+        public async Task<IActionResult> DeleteAsync(
+            [FromRoute] int id)
         {
+            _permissionNeeded.Add("DeletaPessoa");
+            if (!ValidatePermission(_permissionUser, _permissionNeeded))
+                return Forbidden();
             var result = await _personService.DeleteAsync(id);
 
             if (result.IsSuccess)
